requirements.txt
streamlit
pymupdf
pypdf2
pillow
sentence-transformers
faiss-cpu
scikit-learn
spacy
camelot-py[cv]
pytesseract
pandas
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-groq
google-generativeai
